requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Backtesting
zipline-reloaded>=3.0.0
quantstats>=0.0.62
pyfolio>=0.9.2

# LLM and API
openai>=1.0.0  # For OpenRouter API compatibility
requests>=2.31.0
tenacity>=8.2.0

# Data handling
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
